+++ test/rules/lines-around-class.js
@@ -180,5 +180,46 @@ module.exports = extendToClassExpression({
 			`Amount of newlines before class should be 2, but 1 given.`,
 			`Amount of newlines after class should be 2, but 1 given.`
 		)
+	}, {
+		options: [{
+			before: 2,
+			after: 2
+		}],
+		code: concat(
+			`goog.provide('Klass');`,
+			``,
+			``,
+			`// Comment`,
+			``,
+			`/**`,
+			` * @abstract`,
+			` */`,
+			`class Klass {}`,
+			``,
+			`// Comment`,
+			``,
+			``
+		),
+		output: concat(
+			`goog.provide('Klass');`,
+			``,
+			``,
+			`// Comment`,
+			``,
+			``,
+			`/**`,
+			` * @abstract`,
+			` */`,
+			`class Klass {}`,
+			``,
+			``,
+			`// Comment`,
+			``,
+			``
+		),
+		errors: errors(
+			`Amount of newlines before class should be 2, but 1 given.`,
+			`Amount of newlines after class should be 2, but 1 given.`
+		)
 	}]
 });