+++ test/rules/lines-between-props.js
@@ -152,5 +152,58 @@ module.exports = extendToClassExpression({
 			`Unexpected code between props.`,
 			`Amount of newlines between props should be 2, but 1 given.`
 		)
+	}, {
+		options: [{
+			amount: 2
+		}],
+		code: concat(
+			`class Klass {`,
+			`   constructor() {`,
+			`       /**`,
+			`        * @type {number}`,
+			`        */`,
+			`       this.prop;`,
+			``,
+			`       // Comment`,
+			``,
+			`       /**`,
+			`        * @type {number}`,
+			`        */`,
+			`       this.prop = 1;`,
+			``,
+			`       /**`,
+			`        * @const {number}`,
+			`        */`,
+			`       this.CONST = 1;`,
+			`   }`,
+			`}`
+		),
+		output: concat(
+			`class Klass {`,
+			`   constructor() {`,
+			`       /**`,
+			`        * @type {number}`,
+			`        */`,
+			`       this.prop;`,
+			``,
+			`       // Comment`,
+			``,
+			`       /**`,
+			`        * @type {number}`,
+			`        */`,
+			`       this.prop = 1;`,
+			``,
+			``,
+			`       /**`,
+			`        * @const {number}`,
+			`        */`,
+			`       this.CONST = 1;`,
+			`   }`,
+			`}`
+		),
+		errors: errors(
+			`Unexpected comments between props.`,
+			`Amount of newlines between props should be 2, but 1 given.`
+		)
 	}]
 });